Allow docstring to use single quotes and to not be displayed
By specifying None for the docstring no docstring is displayed. By
specifying the quotation mark in the docstring the docstring is not
put into triple quotes allowing a docstring in signle quotes.

The implementation puts the responsibility to add the correct quotation
marks to the python side so we can be more flexible.

The line numbers are now counted from the function body if the docstring
is empty.